Cell suppression with synthetic decimal numbers

GaussSuppressionFromData is run and decimal numbers are added to output by a modified (for sparse matrix efficiency) version of SuppressDec.

Usage

GaussSuppressDec(
  data,
  ...,
  output = NULL,
  digits = 9,
  nRep = NULL,
  rmse = pi/3,
  sparseLimit = 500,
  rndSeed = 123,
  runIpf = FALSE,
  eps = 0.01,
  iter = 100,
  mismatchWarning = TRUE,
  whenDuplicatedInner = NULL,
  whenMixedDuplicatedInner = warning
)

Arguments

data

Input daata as a data frame

...

Further parameters to GaussSuppressionFromData

output

NULL (default), "publish", "inner", "publish_inner", or "publish_inner_x" (x also).

digits

Parameter to RoundWhole. Values close to whole numbers will be rounded.

nRep

NULL or an integer. When >1, several decimal numbers will be generated.

rmse

Desired root mean square error of decimal numbers. Variability around the expected, according to the linear model, inner frequencies. The expected frequencies are calculated from the non-suppressed publishable frequencies.

sparseLimit

Limit for the number of rows of a reduced x-matrix within the algorithm. When exceeded, a new sparse algorithm is used.

rndSeed

If non-NULL, a random generator seed to be used locally within the function without affecting the random value stream in R.

runIpf

When TRUE, additional frequencies are generated by iterative proportional fitting using Mipf.

eps

Parameter to Mipf.

iter

Parameter to Mipf.

mismatchWarning

Whether to produce the warning "Mismatch between whole numbers and suppression", when relevant. When nRep>1, all replicates must satisfy the whole number requirement for non-suppressed cells. When mismatchWarning is integer (>0), this will be used as parameter digits to RoundWhole when doing mismatch checking (can be quite low when nRep>1).

whenDuplicatedInner

Function to be called when default output and when cells marked as inner correspond to several input cells (aggregated) since they correspond to published cells.

whenMixedDuplicatedInner

Function to be called in the case above when some inner cells correspond to published cells (aggregated) and some not (not aggregated).

Value

A data frame where inner cells and cells to be published are combined or output according to parameter output.

Author

Øyvind Langrsud

Examples

z1 <- SSBtoolsData("z1")
GaussSuppressDec(z1, 1:2, 3)
#> GaussSuppression_anySum: ..........................................
#>    region hovedint ant primary suppressed     freqDec isPublish isInner
#> 1   Total    Total 596   FALSE      FALSE 596.0000000      TRUE   FALSE
#> 2   Total    annet  72   FALSE      FALSE  72.0000000      TRUE   FALSE
#> 3   Total   arbeid  52   FALSE      FALSE  52.0000000      TRUE   FALSE
#> 4   Total soshjelp 283   FALSE      FALSE 283.0000000      TRUE   FALSE
#> 5   Total    trygd 189   FALSE      FALSE 189.0000000      TRUE   FALSE
#> 6       A    Total 113   FALSE      FALSE 113.0000000      TRUE   FALSE
#> 7       A    annet  11   FALSE      FALSE  11.0000000      TRUE    TRUE
#> 8       A   arbeid  11   FALSE      FALSE  11.0000000      TRUE    TRUE
#> 9       A soshjelp  55   FALSE      FALSE  55.0000000      TRUE    TRUE
#> 10      A    trygd  36   FALSE      FALSE  36.0000000      TRUE    TRUE
#> 11      B    Total  55   FALSE      FALSE  55.0000000      TRUE   FALSE
#> 12      B    annet   7   FALSE       TRUE   7.0599497      TRUE    TRUE
#> 13      B   arbeid   1    TRUE       TRUE   0.9400503      TRUE    TRUE
#> 14      B soshjelp  29   FALSE      FALSE  29.0000000      TRUE    TRUE
#> 15      B    trygd  18   FALSE      FALSE  18.0000000      TRUE    TRUE
#> 16      C    Total  73   FALSE      FALSE  73.0000000      TRUE   FALSE
#> 17      C    annet   5   FALSE      FALSE   5.0000000      TRUE    TRUE
#> 18      C   arbeid   8   FALSE      FALSE   8.0000000      TRUE    TRUE
#> 19      C soshjelp  35   FALSE      FALSE  35.0000000      TRUE    TRUE
#> 20      C    trygd  25   FALSE      FALSE  25.0000000      TRUE    TRUE
#> 21      D    Total  45   FALSE      FALSE  45.0000000      TRUE   FALSE
#> 22      D    annet  13   FALSE       TRUE  11.4042409      TRUE    TRUE
#> 23      D   arbeid   2    TRUE       TRUE   3.5957591      TRUE    TRUE
#> 24      D soshjelp  17   FALSE      FALSE  17.0000000      TRUE    TRUE
#> 25      D    trygd  13   FALSE      FALSE  13.0000000      TRUE    TRUE
#> 26      E    Total 138   FALSE      FALSE 138.0000000      TRUE   FALSE
#> 27      E    annet   9   FALSE      FALSE   9.0000000      TRUE    TRUE
#> 28      E   arbeid  14   FALSE      FALSE  14.0000000      TRUE    TRUE
#> 29      E soshjelp  63   FALSE      FALSE  63.0000000      TRUE    TRUE
#> 30      E    trygd  52   FALSE      FALSE  52.0000000      TRUE    TRUE
#> 31      F    Total  67   FALSE      FALSE  67.0000000      TRUE   FALSE
#> 32      F    annet  12   FALSE      FALSE  12.0000000      TRUE    TRUE
#> 33      F   arbeid   9   FALSE      FALSE   9.0000000      TRUE    TRUE
#> 34      F soshjelp  24   FALSE      FALSE  24.0000000      TRUE    TRUE
#> 35      F    trygd  22   FALSE      FALSE  22.0000000      TRUE    TRUE
#> 36      G    Total  40   FALSE      FALSE  40.0000000      TRUE   FALSE
#> 37      G    annet   6   FALSE      FALSE   6.0000000      TRUE    TRUE
#> 38      G   arbeid   4   FALSE      FALSE   4.0000000      TRUE    TRUE
#> 39      G soshjelp  22   FALSE      FALSE  22.0000000      TRUE    TRUE
#> 40      G    trygd   8   FALSE      FALSE   8.0000000      TRUE    TRUE
#> 41      H    Total  65   FALSE      FALSE  65.0000000      TRUE   FALSE
#> 42      H    annet   9   FALSE       TRUE  10.5358094      TRUE    TRUE
#> 43      H   arbeid   3    TRUE       TRUE   1.4641906      TRUE    TRUE
#> 44      H soshjelp  38   FALSE      FALSE  38.0000000      TRUE    TRUE
#> 45      H    trygd  15   FALSE      FALSE  15.0000000      TRUE    TRUE
GaussSuppressDec(z1, freqVar = "ant", formula = ~ region + hovedint, maxN = 10)
#>    region hovedint ant primary suppressed     freqDec isPublish isInner
#> 1   Total    Total 596   FALSE      FALSE 596.0000000      TRUE   FALSE
#> 2       A    Total 113   FALSE      FALSE 113.0000000      TRUE   FALSE
#> 3       B    Total  55   FALSE      FALSE  55.0000000      TRUE   FALSE
#> 4       C    Total  73   FALSE      FALSE  73.0000000      TRUE   FALSE
#> 5       D    Total  45   FALSE      FALSE  45.0000000      TRUE   FALSE
#> 6       E    Total 138   FALSE      FALSE 138.0000000      TRUE   FALSE
#> 7       F    Total  67   FALSE      FALSE  67.0000000      TRUE   FALSE
#> 8       G    Total  40   FALSE      FALSE  40.0000000      TRUE   FALSE
#> 9       H    Total  65   FALSE      FALSE  65.0000000      TRUE   FALSE
#> 10  Total    annet  72   FALSE      FALSE  72.0000000      TRUE   FALSE
#> 11  Total   arbeid  52   FALSE      FALSE  52.0000000      TRUE   FALSE
#> 12  Total soshjelp 283   FALSE      FALSE 283.0000000      TRUE   FALSE
#> 13  Total    trygd 189   FALSE      FALSE 189.0000000      TRUE   FALSE
#> 14      A    annet  11      NA         NA  17.9453581     FALSE    TRUE
#> 15      B    annet   7      NA         NA   4.9227606     FALSE    TRUE
#> 16      C    annet   5      NA         NA   8.9060149     FALSE    TRUE
#> 17      D    annet  13      NA         NA   1.3041905     FALSE    TRUE
#> 18      E    annet   9      NA         NA  25.2528144     FALSE    TRUE
#> 19      F    annet  12      NA         NA   8.1267003     FALSE    TRUE
#> 20      G    annet   6      NA         NA   0.8720472     FALSE    TRUE
#> 21      H    annet   9      NA         NA   4.6701139     FALSE    TRUE
#> 22      A   arbeid  11      NA         NA  15.2162489     FALSE    TRUE
#> 23      B   arbeid   1      NA         NA   2.0704964     FALSE    TRUE
#> 24      C   arbeid   8      NA         NA   5.8890847     FALSE    TRUE
#> 25      D   arbeid   2      NA         NA  -0.8504046     FALSE    TRUE
#> 26      E   arbeid  14      NA         NA  23.0735880     FALSE    TRUE
#> 27      F   arbeid   9      NA         NA   3.3548431     FALSE    TRUE
#> 28      G   arbeid   4      NA         NA  -3.0876547     FALSE    TRUE
#> 29      H   arbeid   3      NA         NA   6.3337982     FALSE    TRUE
#> 30      A soshjelp  55      NA         NA  46.8472267     FALSE    TRUE
#> 31      B soshjelp  29      NA         NA  29.9619888     FALSE    TRUE
#> 32      C soshjelp  35      NA         NA  35.1602279     FALSE    TRUE
#> 33      D soshjelp  17      NA         NA  27.9924550     FALSE    TRUE
#> 34      E soshjelp  63      NA         NA  51.0374471     FALSE    TRUE
#> 35      F soshjelp  24      NA         NA  32.8942077     FALSE    TRUE
#> 36      G soshjelp  22      NA         NA  26.2561351     FALSE    TRUE
#> 37      H soshjelp  38      NA         NA  32.8503116     FALSE    TRUE
#> 38      A    trygd  36      NA         NA  32.9911662     FALSE    TRUE
#> 39      B    trygd  18      NA         NA  18.0447542     FALSE    TRUE
#> 40      C    trygd  25      NA         NA  23.0446724     FALSE    TRUE
#> 41      D    trygd  13      NA         NA  16.5537590     FALSE    TRUE
#> 42      E    trygd  52      NA         NA  38.6361506     FALSE    TRUE
#> 43      F    trygd  22      NA         NA  22.6242488     FALSE    TRUE
#> 44      G    trygd   8      NA         NA  15.9594725     FALSE    TRUE
#> 45      H    trygd  15      NA         NA  21.1457762     FALSE    TRUE