Identical to ´NContributorsRule´

The function is included for compatibility after changing the name to NContributorsRule

Usage

NcontributorsRule(
  data,
  freq,
  numVar,
  x,
  maxN = 3,
  protectZeros = FALSE,
  charVar = NULL,
  removeCodes = character(0),
  remove0 = TRUE,
  ...
)

Arguments

data

Input data as a data frame

freq

Vector of aggregate frequencies

numVar

Numerical variables. When several variables, only first is used.

x

Model matrix generated by parent function

maxN

Suppression threshold. Cells where the number of unique contributors (based on charVar) is less than or equal to maxN are marked as primary suppressed. Can be specified as a single numeric value, or as a named list or named vector. When named, the value matching the charVar name will be used. If charVar contains multiple variables and you want different thresholds for each, maxN must be a named list or vector with one value per variable. For example: maxN = list(char1 = 3, char2 = 2).

protectZeros

Suppression parameter. Only TRUE (default) is used implemented.

charVar

Variable(s) with contributor codes. When empty, unique contributor in each row is assumed. When several variables, see details.

removeCodes

Codes to exclude when counting contributors. Can be specified as a character vector (applied to all charVar variables), or as a named list of vectors to use different codes per variable. When using a list, its names must match the variables in charVar. If charVar is empty, codes are interpreted as row indices and converted to integers.

remove0

When set to TRUE (default), data rows in which the first numVar (if any) is zero are excluded from the count of contributors. Alternatively, remove0 can be specified as one or more variable names. In this case, all data rows with a zero in any of the specified variables are omitted from the contributor count. Specifying remove0 as variable name(s) is useful for avoiding warning when there are multiple numVar variables.

...

unused parameters