Default primary function

Function for GaussSuppressionFromData

Usage

PrimaryDefault(freq, maxN = 3, protectZeros = TRUE, freqVar, ...)

Arguments

freq

Vector of output frequencies

maxN

Cells with frequency <= maxN are set as primary suppressed. Can also be a named list or vector, where the value corresponding to freqVar will be used if available. If not found, the name "freq" is tried as an alternative.

protectZeros

When TRUE, cells with zero frequency are set as primary suppressed.

freqVar

Character string used to select the appropriate value from maxN if it is a named object. see maxN above.

...

Unused parameters

Value

primary, GaussSuppression input